Implement plant selection and resolution logic in CLI
- Introduced `GetPlantBySelectorQuery` and `GetPlantBySelectorHandler` to handle plant retrieval based on a selector.
- Added `PlantSelector` utility for resolving plants, including handling ambiguous selectors and not found exceptions.
- Refactored existing plant command handlers to utilize the new plant selection logic, improving code reuse and maintainability.
- Removed outdated plant handling logic from CLI features, streamlining the codebase.
