+++ src/GitForest.Application/Features/Plants/Commands/PlantLifecycleCommands.cs
@@ -1,3 +1,4 @@
+using GitForest.Application.Features.Plants;
 using GitForest.Core;
 using GitForest.Core.Persistence;
 using GitForest.Core.Specifications.Plants;
@@ -50,7 +51,7 @@ internal sealed class AssignPlanterToPlantHandler
         if (request is null)
             throw new ArgumentNullException(nameof(request));
 
-        var resolved = await ResolvePlantAsync(request.Selector, cancellationToken);
+        var resolved = await PlantSelector.ResolveAsync(_plants, request.Selector, cancellationToken);
         var updated = Clone(resolved);
 
         var normalizedPlanterId = (request.PlanterId ?? string.Empty).Trim();
@@ -83,120 +84,6 @@ internal sealed class AssignPlanterToPlantHandler
         }
 
         return updated;
-    }
-
-    private async Task<Plant> ResolvePlantAsync(
-        string selector,
-        CancellationToken cancellationToken
-    )
-    {
-        var sel = (selector ?? string.Empty).Trim();
-        if (sel.Length == 0)
-        {
-            throw new PlantNotFoundException(selector ?? string.Empty);
-        }
-
-        var plants = await _plants.ListAsync(new AllPlantsSpec(), cancellationToken);
-        var matches = FindMatches(plants, sel);
-        if (matches.Count == 1)
-        {
-            return matches[0];
-        }
-
-        if (matches.Count == 0)
-        {
-            throw new PlantNotFoundException(selector ?? string.Empty);
-        }
-
-        throw new PlantAmbiguousSelectorException(
-            selector ?? string.Empty,
-            matches.Select(p => p.Key).ToArray()
-        );
-    }
-
-    internal static IReadOnlyList<Plant> FindMatches(IReadOnlyList<Plant> plants, string selector)
-    {
-        var sel = (selector ?? string.Empty).Trim();
-        if (sel.Length == 0 || plants.Count == 0)
-        {
-            return Array.Empty<Plant>();
-        }
-
-        // 1) Exact key match: <plan-id>:<slug>
-        var exact = plants
-            .Where(p => string.Equals(p.Key, sel, StringComparison.OrdinalIgnoreCase))
-            .ToArray();
-        if (exact.Length > 0)
-        {
-            return exact;
-        }
-
-        // 2) Pxx style stable index into ordered list (best-effort; deterministic ordering).
-        if (TryParsePIndex(sel, out var index))
-        {
-            var ordered = plants.OrderBy(p => p.Key, StringComparer.OrdinalIgnoreCase).ToArray();
-            if (index >= 0 && index < ordered.Length)
-            {
-                return new[] { ordered[index] };
-            }
-
-            return Array.Empty<Plant>();
-        }
-
-        // 3) Slug match: match any plant whose key right-side equals selector.
-        var slugMatches = plants
-            .Where(p =>
-            {
-                var key = p.Key ?? string.Empty;
-                var idx = key.IndexOf(':', StringComparison.Ordinal);
-                if (idx < 0 || idx == key.Length - 1)
-                {
-                    return false;
-                }
-
-                var slug = key[(idx + 1)..];
-                return string.Equals(slug, sel, StringComparison.OrdinalIgnoreCase);
-            })
-            .ToArray();
-
-        return slugMatches;
-    }
-
-    private static bool TryParsePIndex(string selector, out int index)
-    {
-        // Accept P01, p1, P0003 → 1-based ordinal; convert to 0-based index.
-        index = -1;
-        if (selector.Length < 2)
-        {
-            return false;
-        }
-
-        if (selector[0] != 'p' && selector[0] != 'P')
-        {
-            return false;
-        }
-
-        var digits = selector[1..].Trim();
-        if (digits.Length == 0)
-        {
-            return false;
-        }
-
-        foreach (var ch in digits)
-        {
-            if (!char.IsDigit(ch))
-            {
-                return false;
-            }
-        }
-
-        if (!int.TryParse(digits, out var oneBased) || oneBased <= 0)
-        {
-            return false;
-        }
-
-        index = oneBased - 1;
-        return true;
     }
 
     internal static Plant Clone(Plant source)
@@ -246,7 +133,7 @@ internal sealed class UnassignPlanterFromPlantHandler
         if (request is null)
             throw new ArgumentNullException(nameof(request));
 
-        var resolved = await ResolvePlantAsync(request.Selector, cancellationToken);
+        var resolved = await PlantSelector.ResolveAsync(_plants, request.Selector, cancellationToken);
 
         var updated = AssignPlanterToPlantHandler.Clone(resolved);
 
@@ -273,35 +160,6 @@ internal sealed class UnassignPlanterFromPlantHandler
 
         return updated;
     }
-
-    private async Task<Plant> ResolvePlantAsync(
-        string selector,
-        CancellationToken cancellationToken
-    )
-    {
-        var sel = (selector ?? string.Empty).Trim();
-        if (sel.Length == 0)
-        {
-            throw new PlantNotFoundException(selector ?? string.Empty);
-        }
-
-        var plants = await _plants.ListAsync(new AllPlantsSpec(), cancellationToken);
-        var matches = AssignPlanterToPlantHandler.FindMatches(plants, sel);
-        if (matches.Count == 1)
-        {
-            return matches[0];
-        }
-
-        if (matches.Count == 0)
-        {
-            throw new PlantNotFoundException(selector ?? string.Empty);
-        }
-
-        throw new PlantAmbiguousSelectorException(
-            selector ?? string.Empty,
-            matches.Select(p => p.Key).ToArray()
-        );
-    }
 }
 
 public sealed record HarvestPlantCommand(string Selector, bool Force, bool DryRun)
@@ -324,7 +182,7 @@ internal sealed class HarvestPlantHandler : IRequestHandler<HarvestPlantCommand,
         if (request is null)
             throw new ArgumentNullException(nameof(request));
 
-        var resolved = await ResolvePlantAsync(request.Selector, cancellationToken);
+        var resolved = await PlantSelector.ResolveAsync(_plants, request.Selector, cancellationToken);
         var updated = AssignPlanterToPlantHandler.Clone(resolved);
 
         if (
@@ -346,35 +204,6 @@ internal sealed class HarvestPlantHandler : IRequestHandler<HarvestPlantCommand,
         }
 
         return updated;
-    }
-
-    private async Task<Plant> ResolvePlantAsync(
-        string selector,
-        CancellationToken cancellationToken
-    )
-    {
-        var sel = (selector ?? string.Empty).Trim();
-        if (sel.Length == 0)
-        {
-            throw new PlantNotFoundException(selector ?? string.Empty);
-        }
-
-        var plants = await _plants.ListAsync(new AllPlantsSpec(), cancellationToken);
-        var matches = AssignPlanterToPlantHandler.FindMatches(plants, sel);
-        if (matches.Count == 1)
-        {
-            return matches[0];
-        }
-
-        if (matches.Count == 0)
-        {
-            throw new PlantNotFoundException(selector ?? string.Empty);
-        }
-
-        throw new PlantAmbiguousSelectorException(
-            selector ?? string.Empty,
-            matches.Select(p => p.Key).ToArray()
-        );
     }
 }
 
@@ -398,7 +227,7 @@ internal sealed class ArchivePlantHandler : IRequestHandler<ArchivePlantCommand,
         if (request is null)
             throw new ArgumentNullException(nameof(request));
 
-        var resolved = await ResolvePlantAsync(request.Selector, cancellationToken);
+        var resolved = await PlantSelector.ResolveAsync(_plants, request.Selector, cancellationToken);
         var updated = AssignPlanterToPlantHandler.Clone(resolved);
 
         if (
@@ -421,35 +250,6 @@ internal sealed class ArchivePlantHandler : IRequestHandler<ArchivePlantCommand,
 
         return updated;
     }
-
-    private async Task<Plant> ResolvePlantAsync(
-        string selector,
-        CancellationToken cancellationToken
-    )
-    {
-        var sel = (selector ?? string.Empty).Trim();
-        if (sel.Length == 0)
-        {
-            throw new PlantNotFoundException(selector ?? string.Empty);
-        }
-
-        var plants = await _plants.ListAsync(new AllPlantsSpec(), cancellationToken);
-        var matches = AssignPlanterToPlantHandler.FindMatches(plants, sel);
-        if (matches.Count == 1)
-        {
-            return matches[0];
-        }
-
-        if (matches.Count == 0)
-        {
-            throw new PlantNotFoundException(selector ?? string.Empty);
-        }
-
-        throw new PlantAmbiguousSelectorException(
-            selector ?? string.Empty,
-            matches.Select(p => p.Key).ToArray()
-        );
-    }
 }
 
 public sealed record RemovePlantCommand(string Selector, bool Force, bool DryRun) : IRequest<Plant>;
@@ -468,7 +268,7 @@ internal sealed class RemovePlantHandler : IRequestHandler<RemovePlantCommand, P
         if (request is null)
             throw new ArgumentNullException(nameof(request));
 
-        var resolved = await ResolvePlantAsync(request.Selector, cancellationToken);
+        var resolved = await PlantSelector.ResolveAsync(_plants, request.Selector, cancellationToken);
 
         if (
             !request.Force
@@ -488,32 +288,6 @@ internal sealed class RemovePlantHandler : IRequestHandler<RemovePlantCommand, P
         }
 
         return snapshot;
-    }
-
-    private async Task<Plant> ResolvePlantAsync(string selector, CancellationToken cancellationToken)
-    {
-        var sel = (selector ?? string.Empty).Trim();
-        if (sel.Length == 0)
-        {
-            throw new PlantNotFoundException(selector ?? string.Empty);
-        }
-
-        var plants = await _plants.ListAsync(new AllPlantsSpec(), cancellationToken);
-        var matches = AssignPlanterToPlantHandler.FindMatches(plants, sel);
-        if (matches.Count == 1)
-        {
-            return matches[0];
-        }
-
-        if (matches.Count == 0)
-        {
-            throw new PlantNotFoundException(selector ?? string.Empty);
-        }
-
-        throw new PlantAmbiguousSelectorException(
-            selector ?? string.Empty,
-            matches.Select(p => p.Key).ToArray()
-        );
     }
 }
