+++ src/GitForest.Cli/Commands/PlannersCommand.cs
@@ -68,19 +68,7 @@ public static class PlannersCommand
                 }
                 catch (ForestStore.ForestNotInitializedException)
                 {
-                    if (output.Json)
-                    {
-                        output.WriteJsonError(
-                            code: "forest_not_initialized",
-                            message: "Forest not initialized"
-                        );
-                    }
-                    else
-                    {
-                        output.WriteErrorLine("Error: forest not initialized");
-                    }
-
-                    return ExitCodes.ForestNotInitialized;
+                    return BaseCommand.WriteForestNotInitialized(output);
                 }
             }
         );
@@ -131,7 +119,7 @@ public static class PlannersCommand
                     && string.IsNullOrWhiteSpace(plannerId)
                 )
                 {
-                    return WriteInvalidArguments(
+                    return BaseCommand.WriteInvalidArguments(
                         output,
                         "Specify --all, --plan, or --planner",
                         new { all, planId, plannerId }
@@ -209,36 +197,11 @@ public static class PlannersCommand
                 }
                 catch (ForestStore.ForestNotInitializedException)
                 {
-                    if (output.Json)
-                    {
-                        output.WriteJsonError(
-                            code: "forest_not_initialized",
-                            message: "Forest not initialized"
-                        );
-                    }
-                    else
-                    {
-                        output.WriteErrorLine("Error: forest not initialized");
-                    }
-
-                    return ExitCodes.ForestNotInitialized;
+                    return BaseCommand.WriteForestNotInitialized(output);
                 }
                 catch (AppPlans.PlanNotInstalledException)
                 {
-                    if (output.Json)
-                    {
-                        output.WriteJsonError(
-                            code: "plan_not_found",
-                            message: "Plan not found",
-                            details: new { planId }
-                        );
-                    }
-                    else
-                    {
-                        output.WriteErrorLine($"Error: plan not found: {planId}");
-                    }
-
-                    return ExitCodes.PlanNotFound;
+                    return BaseCommand.WritePlanNotFound(output, planId ?? string.Empty);
                 }
             }
         );
@@ -259,21 +222,4 @@ public static class PlannersCommand
         return value.Length <= max ? value : value[..Math.Max(0, max - 3)] + "...";
     }
 
-    private static int WriteInvalidArguments(Output output, string message, object? details)
-    {
-        if (output.Json)
-        {
-            output.WriteJsonError(
-                code: "invalid_arguments",
-                message: message,
-                details: details
-            );
-        }
-        else
-        {
-            output.WriteErrorLine($"Error: {message}");
-        }
-
-        return ExitCodes.InvalidArguments;
-    }
 }
