+++ src/GitForest.Cli/Commands/PlansCommand.cs
@@ -1,5 +1,5 @@
-using GitForest.Mediator;
 using System.CommandLine;
+using GitForest.Mediator;
 using AppPlans = GitForest.Application.Features.Plans;
 
 namespace GitForest.Cli.Commands;
@@ -8,193 +8,175 @@ public static class PlansCommand
 {
     public static Command Build(CliOptions cliOptions, IMediator mediator)
     {
-        var plansCommand = new Command("plans", "Manage plans");
+        return CliCommandBuilder
+            .Create("plans", "Manage plans")
+            .Subcommand("list", "List installed plans", ConfigureListCommand)
+            .Subcommand("install", "Install a plan", ConfigureInstallCommand)
+            .Build();
 
-        var listCommand = new Command("list", "List installed plans");
-        listCommand.SetAction(
-            async (parseResult, token) =>
-            {
-                var output = parseResult.GetOutput(cliOptions);
-
-                try
+        void ConfigureListCommand(CliCommandBuilder listCommand)
+        {
+            listCommand.Action(
+                async (parseResult, token) =>
                 {
-                    var forestDir = ForestStore.GetDefaultForestDir();
-                    ForestStore.EnsureInitialized(forestDir);
+                    var output = parseResult.GetOutput(cliOptions);
 
-                    var plans = await mediator.Send(new AppPlans.ListPlansQuery(), token);
-                    if (output.Json)
+                    try
                     {
-                        output.WriteJson(
-                            new
-                            {
-                                plans = plans
-                                    .Select(p => new
-                                    {
-                                        id = p.Id,
-                                        name = p.Id,
-                                        version = p.Version,
-                                        category = string.Empty,
-                                        author = p.Author,
-                                        license = p.License,
-                                        repository = p.Repository,
-                                        homepage = p.Homepage,
-                                        source = p.Source,
-                                        installedAt = p.InstalledDate == default
-                                            ? string.Empty
-                                            : p.InstalledDate.ToString("O"),
-                                        sha256 = string.Empty,
-                                    })
-                                    .ToArray(),
-                            }
-                        );
-                    }
-                    else
-                    {
-                        if (plans.Count == 0)
+                        var forestDir = ForestStore.GetDefaultForestDir();
+                        ForestStore.EnsureInitialized(forestDir);
+
+                        var plans = await mediator.Send(new AppPlans.ListPlansQuery(), token);
+                        if (output.Json)
                         {
-                            output.WriteLine("No plans installed");
+                            output.WriteJson(
+                                new
+                                {
+                                    plans = plans
+                                        .Select(p => new
+                                        {
+                                            id = p.Id,
+                                            name = p.Id,
+                                            version = p.Version,
+                                            category = string.Empty,
+                                            author = p.Author,
+                                            license = p.License,
+                                            repository = p.Repository,
+                                            homepage = p.Homepage,
+                                            source = p.Source,
+                                            installedAt = p.InstalledDate == default
+                                                ? string.Empty
+                                                : p.InstalledDate.ToString("O"),
+                                            sha256 = string.Empty,
+                                        })
+                                        .ToArray(),
+                                }
+                            );
                         }
                         else
                         {
-                            output.WriteLine(
-                                $"{PadRight("Id", 28)} {PadRight("Version", 10)} {PadRight("Category", 20)} {PadRight("InstalledAt", 20)}"
-                            );
-
-                            foreach (var plan in plans)
+                            if (plans.Count == 0)
                             {
-                                var version = string.IsNullOrWhiteSpace(plan.Version)
-                                    ? "-"
-                                    : plan.Version;
-                                var category = "-";
-                                var installedAt =
-                                    plan.InstalledDate == default
-                                        ? "-"
-                                        : FormatInstalledAt(plan.InstalledDate.ToString("O"));
-
+                                output.WriteLine("No plans installed");
+                            }
+                            else
+                            {
                                 output.WriteLine(
-                                    $"{PadRight(plan.Id, 28)} {PadRight(version, 10)} {PadRight(Truncate(category, 20), 20)} {PadRight(installedAt, 20)}"
+                                    $"{PadRight("Id", 28)} {PadRight("Version", 10)} {PadRight("Category", 20)} {PadRight("InstalledAt", 20)}"
                                 );
+
+                                foreach (var plan in plans)
+                                {
+                                    var version = string.IsNullOrWhiteSpace(plan.Version)
+                                        ? "-"
+                                        : plan.Version;
+                                    var category = "-";
+                                    var installedAt =
+                                        plan.InstalledDate == default
+                                            ? "-"
+                                            : FormatInstalledAt(plan.InstalledDate.ToString("O"));
+
+                                    output.WriteLine(
+                                        $"{PadRight(plan.Id, 28)} {PadRight(version, 10)} {PadRight(Truncate(category, 20), 20)} {PadRight(installedAt, 20)}"
+                                    );
+                                }
                             }
                         }
-                    }
 
-                    return ExitCodes.Success;
-                }
-                catch (ForestStore.ForestNotInitializedException)
-                {
-                    if (output.Json)
+                        return ExitCodes.Success;
+                    }
+                    catch (ForestStore.ForestNotInitializedException)
                     {
-                        output.WriteJsonError(
-                            code: "forest_not_initialized",
-                            message: "Forest not initialized"
-                        );
+                        return BaseCommand.WriteForestNotInitialized(output);
                     }
-                    else
-                    {
-                        output.WriteErrorLine("Error: forest not initialized");
-                    }
-
-                    return ExitCodes.ForestNotInitialized;
                 }
-            }
-        );
-        plansCommand.Subcommands.Add(listCommand);
+            );
+        }
 
-        var installCommand = new Command("install", "Install a plan");
-        var sourceArg = new Argument<string>("source")
+        void ConfigureInstallCommand(CliCommandBuilder installCommand)
         {
-            Description = "Plan source (GitHub slug, URL, or local path)",
-        };
-        installCommand.Arguments.Add(sourceArg);
-        installCommand.SetAction(
-            async (parseResult, token) =>
+            var sourceArg = new Argument<string>("source")
             {
-                var output = parseResult.GetOutput(cliOptions);
-                var source = parseResult.GetRequiredValue(sourceArg);
+                Description = "Plan source (GitHub slug, URL, or local path)",
+            };
 
-                try
-                {
-                    var forestDir = ForestStore.GetDefaultForestDir();
-                    ForestStore.EnsureInitialized(forestDir);
-
-                    var installed = await mediator.Send(
-                        new AppPlans.InstallPlanCommand(Source: source),
-                        token
-                    );
-                    if (output.Json)
+            installCommand
+                .AddArgument(sourceArg)
+                .Action(
+                    async (parseResult, token) =>
                     {
-                        output.WriteJson(
-                            new
+                        var output = parseResult.GetOutput(cliOptions);
+                        var source = parseResult.GetRequiredValue(sourceArg);
+
+                        try
+                        {
+                            var forestDir = ForestStore.GetDefaultForestDir();
+                            ForestStore.EnsureInitialized(forestDir);
+
+                            var installed = await mediator.Send(
+                                new AppPlans.InstallPlanCommand(Source: source),
+                                token
+                            );
+                            if (output.Json)
                             {
-                                status = "installed",
-                                source,
-                                planId = installed.Id,
-                                version = installed.Version,
+                                output.WriteJson(
+                                    new
+                                    {
+                                        status = "installed",
+                                        source,
+                                        planId = installed.Id,
+                                        version = installed.Version,
+                                    }
+                                );
                             }
-                        );
-                    }
-                    else
-                    {
-                        output.WriteLine($"Installed plan from: {source}");
-                    }
+                            else
+                            {
+                                output.WriteLine($"Installed plan from: {source}");
+                            }
 
-                    return ExitCodes.Success;
-                }
-                catch (ForestStore.ForestNotInitializedException)
-                {
-                    if (output.Json)
-                    {
-                        output.WriteJsonError(
-                            code: "forest_not_initialized",
-                            message: "Forest not initialized"
-                        );
-                    }
-                    else
-                    {
-                        output.WriteErrorLine("Error: forest not initialized");
-                    }
+                            return ExitCodes.Success;
+                        }
+                        catch (ForestStore.ForestNotInitializedException)
+                        {
+                            return BaseCommand.WriteForestNotInitialized(output);
+                        }
+                        catch (FileNotFoundException)
+                        {
+                            if (output.Json)
+                            {
+                                output.WriteJsonError(
+                                    code: "plan_not_found",
+                                    message: "Plan file not found",
+                                    details: new { source }
+                                );
+                            }
+                            else
+                            {
+                                output.WriteErrorLine("Error: plan file not found");
+                            }
 
-                    return ExitCodes.ForestNotInitialized;
-                }
-                catch (FileNotFoundException)
-                {
-                    if (output.Json)
-                    {
-                        output.WriteJsonError(
-                            code: "plan_not_found",
-                            message: "Plan file not found",
-                            details: new { source }
-                        );
-                    }
-                    else
-                    {
-                        output.WriteErrorLine("Error: plan file not found");
-                    }
+                            return ExitCodes.PlanNotFound;
+                        }
+                        catch (InvalidDataException ex)
+                        {
+                            if (output.Json)
+                            {
+                                output.WriteJsonError(
+                                    code: "schema_validation_failed",
+                                    message: "Invalid plan YAML",
+                                    details: new { error = ex.Message }
+                                );
+                            }
+                            else
+                            {
+                                output.WriteErrorLine("Error: invalid plan YAML");
+                            }
 
-                    return ExitCodes.PlanNotFound;
-                }
-                catch (InvalidDataException ex)
-                {
-                    if (output.Json)
-                    {
-                        output.WriteJsonError(
-                            code: "schema_validation_failed",
-                            message: "Invalid plan YAML",
-                            details: new { error = ex.Message }
-                        );
+                            return ExitCodes.SchemaValidationFailed;
+                        }
                     }
-                    else
-                    {
-                        output.WriteErrorLine("Error: invalid plan YAML");
-                    }
-
-                    return ExitCodes.SchemaValidationFailed;
-                }
-            }
-        );
-        plansCommand.Subcommands.Add(installCommand);
-
-        return plansCommand;
+                );
+        }
     }
 
     private static string PadRight(string value, int width)
